airflow/requirements.txt
google-cloud
pandas